requirements.txt
langchain-cohere>=0.1.0
langchain-google-genai>=0.1.0
langchain-mistralai>=0.1.0
langchain-perplexity>=0.1.0
langchain-huggingface>=0.1.0
cohere>=4.50.0
google-generativeai>=0.4.1
mistralai>=0.1.0

huggingface_hub>=0.23.0
streamlit>=1.32.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.0
langchain-openai>=0.1.0
langchain-anthropic>=0.1.0
langchain-together>=0.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
opencv-python-headless>=4.8.0.74
requests>=2.31.0
pydantic>=2.5.0